+++ src/resources/stacks/tasks/stack.ts
@@ -2,20 +2,7 @@ import * as Request from "../../../common/api/request";
 import { links, StandardParams } from "../../../common/api";
 import { ResourceId, Task, CreatedTask } from "../../../common/structs";
 
-export type StackAction = "build" | "prune";
-
-export async function buildStack(
-  params: StandardParams & {
-    id: ResourceId;
-  },
-) {
-  return task({
-    ...params,
-    value: {
-      action: "build",
-    },
-  });
-}
+export type StackAction = "prune";
 
 /** Removes all builds that are not in use and have not been created in the last 30 minutes */
 export async function pruneBuilds(